+++ contracts/proto/Encoder.sol
@@ -6,19 +6,14 @@ library Encoder {
     uint64 private constant _MAX_UINT64 = 0xFFFFFFFFFFFFFFFF;
 
     function cdcEncode(string memory item) internal pure returns (bytes memory) {
-        uint estimatedSize = 1 + ProtoBufRuntime._sz_lendelim(bytes(item).length);
+        uint256 estimatedSize = 1 + ProtoBufRuntime._sz_lendelim(bytes(item).length);
         bytes memory bs = new bytes(estimatedSize);
 
         uint256 offset = 32;
         uint256 pointer = 32;
 
         if (bytes(item).length > 0) {
-            pointer += ProtoBufRuntime._encode_key(
-                1,
-                ProtoBufRuntime.WireType.LengthDelim,
-                pointer,
-                bs
-            );
+            pointer += ProtoBufRuntime._encode_key(1, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
             pointer += ProtoBufRuntime._encode_string(item, pointer, bs);
         }
 
@@ -30,19 +25,14 @@ library Encoder {
     }
 
     function cdcEncode(bytes memory item) internal pure returns (bytes memory) {
-        uint estimatedSize = 1 + ProtoBufRuntime._sz_lendelim(item.length);
+        uint256 estimatedSize = 1 + ProtoBufRuntime._sz_lendelim(item.length);
         bytes memory bs = new bytes(estimatedSize);
 
         uint256 offset = 32;
         uint256 pointer = 32;
 
         if (item.length > 0) {
-            pointer += ProtoBufRuntime._encode_key(
-                1,
-                ProtoBufRuntime.WireType.LengthDelim,
-                pointer,
-                bs
-            );
+            pointer += ProtoBufRuntime._encode_key(1, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
             pointer += ProtoBufRuntime._encode_bytes(item, pointer, bs);
         }
 
@@ -54,19 +44,14 @@ library Encoder {
     }
 
     function cdcEncode(int64 item) internal pure returns (bytes memory) {
-        uint estimatedSize = 1 + ProtoBufRuntime._sz_int64(item);
+        uint256 estimatedSize = 1 + ProtoBufRuntime._sz_int64(item);
         bytes memory bs = new bytes(estimatedSize);
 
         uint256 offset = 32;
         uint256 pointer = 32;
 
         if (item != 0) {
-            pointer += ProtoBufRuntime._encode_key(
-                1,
-                ProtoBufRuntime.WireType.Varint,
-                pointer,
-                bs
-            );
+            pointer += ProtoBufRuntime._encode_key(1, ProtoBufRuntime.WireType.Varint, pointer, bs);
             pointer += ProtoBufRuntime._encode_int64(item, pointer, bs);
         }
